requirements.txt
fastapi
uvicorn
sqlalchemy
proxmoxer
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
jinja2